autopilot/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentspan-autopilot"
version = "0.1.0"
description = "Agentspan Autopilot — autonomous agent orchestration from YAML configs"
readme = "README.md"
license = {text = "MIT License"}
requires-python = ">=3.10"
dependencies = [
    "agentspan",
    "pyyaml>=6.0",
    "httpx>=0.24",
    "trafilatura>=1.6",
    "prompt_toolkit>=3.0",
    "ddgs>=9.0",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.optional-dependencies]
markitdown = ["markitdown>=0.1"]
langextract = ["langextract>=0.1"]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30
markers = [
    "network: marks tests that require network access (deselect with '-m \"not network\"')",
    "e2e: marks end-to-end tests requiring server and LLM API (deselect with '-m \"not e2e\"')",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[tool.uv.sources]
agentspan = {path = "../sdk/python"}
